+++ src/Examples/NetCore/Model.g.fs
@@ -1,5 +1,5 @@
-//55aa31b7-a708-eec1-2660-235eea801516
-//51774502-0567-58bc-8522-022d1851b6c1
+//1f393f24-1629-f7a3-f49d-eaeb02c60d70
+//fc724c27-fcd8-96fc-37e4-0a3ecccc3b9d
 #nowarn "49" // upper case patterns
 #nowarn "66" // upcast is unncecessary
 #nowarn "1337" // internal types
@@ -19,11 +19,7 @@ type AdaptiveMyModel(value : MyModel) =
             o
         Adaptify.FSharp.Core.AdaptiveResult<Model.MyModel, Model.AdaptiveMyModel, Model.AdaptiveMyModel, Microsoft.FSharp.Core.string, Microsoft.FSharp.Core.string, FSharp.Data.Adaptive.aval<Microsoft.FSharp.Core.string>>(value.a, (fun (v : MyModel) -> AdaptiveMyModel(v) :> System.Object), (fun (o : System.Object) (v : MyModel) -> (unbox<AdaptiveMyModel> o).Update(v) :> System.Object), (fun (o : System.Object) -> unbox<AdaptiveMyModel> o), (fun (v : MyModel) -> AdaptiveMyModel(v) :> System.Object), __arg5, (fun (o : System.Object) -> unbox<AdaptiveMyModel> o), (fun (v : Microsoft.FSharp.Core.string) -> v :> System.Object), (fun (o : System.Object) (v : Microsoft.FSharp.Core.string) -> v :> System.Object), (fun (o : System.Object) -> unbox<Microsoft.FSharp.Core.string> o), (fun (v : Microsoft.FSharp.Core.string) -> FSharp.Data.Adaptive.cval(v) :> System.Object), __arg11, (fun (o : System.Object) -> unbox<FSharp.Data.Adaptive.cval<Microsoft.FSharp.Core.string>> o :> FSharp.Data.Adaptive.aval<Microsoft.FSharp.Core.string>))
     let _b_ = FSharp.Data.Adaptive.cval(value.b)
-    let _list_ =
-        let inline __arg2 (m : AdaptiveMyModel) (v : MyModel) =
-            m.Update(v)
-            m
-        FSharp.Data.Traceable.ChangeableModelList(value.list, (fun (v : MyModel) -> AdaptiveMyModel(v)), __arg2, (fun (m : AdaptiveMyModel) -> m))
+    let _list_ = FSharp.Data.Adaptive.cval(value.list)
     let _map_ =
         let inline __arg2 (m : AdaptiveMyModel) (v : MyModel) =
             m.Update(v)
@@ -39,11 +35,11 @@ type AdaptiveMyModel(value : MyModel) =
             __adaptive.MarkOutdated()
             _a_.Update(value.a)
             _b_.Value <- value.b
-            _list_.Update(value.list)
+            _list_.Value <- value.list
             _map_.Update(value.map)
     member __.Current = __adaptive
     member __.a = _a_ :> FSharp.Data.Adaptive.aval<Adaptify.FSharp.Core.AdaptiveResultCase<MyModel, AdaptiveMyModel, AdaptiveMyModel, Microsoft.FSharp.Core.string, Microsoft.FSharp.Core.string, FSharp.Data.Adaptive.aval<Microsoft.FSharp.Core.string>>>
     member __.b = _b_ :> FSharp.Data.Adaptive.aval<Microsoft.FSharp.Core.float>
-    member __.list = _list_ :> FSharp.Data.Adaptive.alist<AdaptiveMyModel>
+    member __.list = _list_ :> FSharp.Data.Adaptive.aval<FSharp.Data.Adaptive.HashSet<MyModel>>
     member __.map = _map_ :> FSharp.Data.Adaptive.amap<Microsoft.FSharp.Core.int, AdaptiveMyModel>