+++ src/Examples/NetCore/Program.fs
@@ -4,12 +4,14 @@ open Adaptify.FSharp.Core
 open Model
 open Adaptify
 
+
+
 [<EntryPoint>]
 let main _argv =
     let test = Unpersist.instance<MyModel, AdaptiveMyModel>
 
-    let v = test.init { a = Error "bad"; b = 3.0; list = IndexList.empty; map = HashMap.empty }
-    test.update v { a = Error "bad"; b = 4.0; list = IndexList.empty; map = HashMap.empty }
+    let v = test.init { a = Error "bad"; b = 3.0; list = HashSet.ofList []; map = HashMap.empty }
+    test.update v { a = Error "bad"; b = 4.0; list = HashSet.empty; map = HashMap.empty }
 
     let _ : aval<MyModel> = v.Current